Remove unnecessary File tests
Remove File tests that were testing if the File gets the lines up to
where the cursor is. That is an implementation detail, and we don’t
need to test it.

We do test that we can generate the tree of lines by indentation level,
which is what we really care about.